refactor(codegen): colocate data copy costs
